src/test/java/com/springpractice/dipractice/controller/ConstructorInjectedControllerTest.java
package com.springpractice.dipractice.controller;

import com.springpractice.dipractice.services.GreetingService;
import com.springpractice.dipractice.services.GreetingServiceImpl;
import org.junit.Before;
import org.junit.Test;

import static org.junit.Assert.assertEquals;

public class ConstructorInjectedControllerTest {

    private ConstructorInjectedController constructorInjectedController;

    @Before
    public void setUp() throws Exception{
        this.constructorInjectedController = new ConstructorInjectedController(new GreetingServiceImpl());
    }

    @Test
    public void testGreeting() throws Exception{
        assertEquals(GreetingServiceImpl.HEY_SPRING, this.constructorInjectedController.sayHello());
    }
}
